Health_Fitness_App/data/requirements.txt
streamlit~=1.39.0
pandas~=2.2.3
python-dotenv~=1.0.1
plotly~=5.24.1
openai~=1.54.0
python-dotenv
bcrypt~=4.1.1